Reorganize buildplanner package
