+++ pkg/platform/api/buildplanner/request/pushcommit.go
@@ -1,7 +1,7 @@
 package request
 
 import (
-	model "github.com/ActiveState/cli/pkg/platform/api/buildplanner"
+	model "github.com/ActiveState/cli/pkg/platform/api/buildplanner/model"
 )
 
 func PushCommit(owner, project, parentCommit, branchRef, description string, script model.BuildExpression) *buildPlanByPushCommit {
