+++ pkg/platform/runtime/store/store.go
@@ -7,7 +7,7 @@ import (
 	"path/filepath"
 	"strings"
 
-	bpModel "github.com/ActiveState/cli/pkg/platform/api/buildplanner"
+	bpModel "github.com/ActiveState/cli/pkg/platform/api/buildplanner/model"
 
 	"github.com/ActiveState/cli/internal/constants"
 	"github.com/ActiveState/cli/internal/errs"
