+++ pkg/platform/runtime/testhelper/testhelper.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 
 	"github.com/ActiveState/cli/internal/environment"
-	model "github.com/ActiveState/cli/pkg/platform/api/buildplanner"
+	model "github.com/ActiveState/cli/pkg/platform/api/buildplanner/model"
 	"github.com/ActiveState/cli/pkg/platform/api/headchef/headchef_models"
 	"github.com/ActiveState/cli/pkg/platform/api/inventory/inventory_models"
 	"github.com/stretchr/testify/require"
